Pretty printing the LDES feed
This is not relevant in 99% of the cases, but it maks it obvious what
has changed later on.  Here for clearing the commits.  LDES feeds
generally don't belong in repositories.

Got the initial LDES feed together with @Yalz who showed we could import
the feed using Digitaal Vlaanderen's toolchain given some changes, and
this pretty and rename is on @madnificent's conto.

Pretty printing through jq and bash:

  for file in *
  do
    cat $file | jq > ../pretty-output/$file
  done