Use relative links instead of absolute links
The provided dumps use absolute links.  We will not be able to use this
during ingestion because they're not fully descirptive of the content we
have available.

We may change URIs due to this, but in discussion with @Yalz we thought
this was fine.

I considered this further and the URIs of our data source will indeed
not match those in ipdc tni anymore but given that these are under tni,
they would never correctly match public if they were defined there.
URIs which are outside of this namespace would still match.  We know
ingestion should happen differently in a real case, so for a one-off
import, this is a sufficiently stable base.

Command is more trivial than the text above:

  sed -i 's@https://ipdc.tni-vlaanderen.be/doc/instantiesnapshot@./instantiesnapshot@g' *