Add UUID generation service
We will need UUIDs if we want to render content in the frontend.  This
is going to be our fastest path to having those.  It runs as a CRON job
in this setting which isn't how it should be wired but it _will_ add
UUIDs for the missing entities at this point and that's good enough for
now.