Steps towards cleaner wiring for LDIO
This is an adaptation to the LDIO configuration which should make it
read without binding to host.  It seems LDIO isn't picking up the new
configuration though.

We could have this boot up always if we can store the sqlite database
too because that would only make it fetch the last page every so often.