Add Ollama infrastructure
This is systems-programming style.  There should likely be one service
named "ollama" which runs locally and which could be considered an
external service also.  Although we're putting the pieces together as to
where and how it should live exactly, this is a starting
point.

This is copied from another repository by @Zed-Is-Dead and @cris-dandre,
prose by @madnificent.